resume_analyzer/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
lxml==6.0.1
pdfminer.six==20250506
pycparser==2.22
python-docx==1.2.0
typing_extensions==4.15.0
